requirements.txt
fastapi
uvicorn
streamlit
pandas
scikit-learn
xgboost
joblib
python-dotenv
matplotlib
requests
imbalanced-learn